Set px and py to small value when both are 0, use long long for current timestamp

+++ P2-Unscented-Kalman-Filter/CarND-Unscented-Kalman-Filter-Project/src/ukf.h
@@ -75,7 +75,7 @@ public:
 
   MatrixXd R_radar_;
 
-  double previous_timestamp_;
+  long long previous_timestamp_;
 
   MatrixXd H_laser_;
   /**
